Apply patch from Erik Hovland to add some missing null pointer checks to the frontend. Some hunks of the patch no longer applied, so not all issues originally identified in the ticket have been fixed, please submit a new patch if required. Closes #7165
git-svn-id: http://svn.mythtv.org/svn/trunk@23945 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythfrontend/customedit.cpp
@@ -794,6 +794,8 @@ void CustomEdit::storeRule(bool is_search, bool is_new)
 void CustomEdit::deleteRule(void)
 {
     MythUIButtonListItem* item = m_clauseList->GetItemCurrent();
+    if (!item)
+        return;
 
     MSqlQuery query(MSqlQuery::InitCon());
     query.prepare("DELETE FROM customexample "
